refactor: extract snapshot writer
Put snapshotwriter into it's own struct and have the hdfs snapshotwriter
embed the snapshot writer to perform the generic portions of the writer

This will allow future table implementations to reuse the writer.
